requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
